reworked the backend to allow easier implementation of forks with different logic vs upstream
