+++ _00_CLI.spec
@@ -5,14 +5,44 @@ block_cipher = None
 
 
 a = Analysis(['_00_CLI.py'],
-             pathex=['chia_blockchain'],
+             pathex=[],
              binaries=[],
              datas=[('icon.ico', '.'),
 			  ('version.txt', '.'),
               ('donation.gif', '.'),
               (sys.prefix+r'/tcl/tix8.4.3', 'tcl/tix8.4.3'),
-              ('chia_blockchain/chia', 'chia')],
-             hiddenimports=[],
+              ('chives_blockchain/chives', 'chives_blockchain/chives'),
+              ('chia_blockchain/chia', 'chia_blockchain/chia')],
+             hiddenimports=['tabulate',
+							'pyinstaller',
+							'yaml',
+							'aiohttp',
+							'aiosqlite',
+							'bitstring',
+							'blspy',
+							'chiabip158',
+							'chiapos',
+							'chiavdf',
+							'click',
+							'clvm',
+							'clvm_rs',
+							'clvm_tools',
+							'colorama',
+							'colorlog',
+							'concurrent-log-handler',
+							'cryptography',
+							'dnspython',
+							'fasteners',
+							'keyrings.cryptfile.cryptfile',
+							'watchdog.events',
+							'watchdog.observers',
+							'websockets',
+							'pillow',
+							'requests',
+							'dataclasses',
+							'clvm_tools.curry',
+							'clvm_tools.clvmc'
+			 ],
              hookspath=[],
              hooksconfig={},
              runtime_hooks=[],
